(6/5) Text fix & Set Cells margin

+++ app/src/main/java/com/eungu/habittraining/MainActivity.java
@@ -44,7 +44,6 @@ public class MainActivity extends Activity {
     private int level = -1, phase = -1, days = -1, rested = -1;
     private UnityPlayer m_UnityPlayer;
     private String now;
-    private boolean isFailed;
 
     AlertDialog.Builder alertDialogBuilder;
 
@@ -54,25 +53,23 @@ public class MainActivity extends Activity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        isFailed = false;
         backPressCloseHandler = new BackPressCloseHandler(this);
 
         InitializeToday();
         MakePassedDataList();
         loadUnityView();
         MakeTabs();
-        SetCheckFailed();
         if(isClear()){
             TextView tv = findViewById(R.id.text_fail);
             tv.setText("모든 목표를 달성하였습니다.\n수고하셨습니다!!");
         }
-        else if(isFailed == false) {
+        else if(isFailed() == false) {
             findViewById(R.id.text_fail).setVisibility(View.GONE);
             MakeGoalList();
         }
         else {
             TextView tv = findViewById(R.id.text_fail);
-            tv.setText(rested + "회 목표를 달성하지 못하였습니다...\n초기화 후 다시 해주세요.");
+            tv.setText("3회이상 목표를 달성하지 못하였습니다...\n초기화 후 다시 해주세요.");
         }
     }
 
@@ -84,8 +81,9 @@ public class MainActivity extends Activity {
         return false;
     }
 
-    private void SetCheckFailed(){
-        if(rested >= 3) isFailed = true;
+    private boolean isFailed(){
+        if(rested >= 3) return true;
+        else return false;
     }
 
     private void InitializeToday(){
@@ -114,7 +112,7 @@ public class MainActivity extends Activity {
         sql = String.format("SELECT * FROM training WHERE today LIKE '%s';", now);
         c = _db.rawQuery(sql, null);
 
-        if(c.getCount() == 0 && isClear() == false){
+        if(c.getCount() == 0 && isClear() == false && isFailed() == false){
             if(DidCompleteYesterday() == false){
                 rested += 1;
                 _db.execSQL("DROP TABLE IF EXISTS grown;");
@@ -221,7 +219,7 @@ public class MainActivity extends Activity {
                 break;
             case 3:
                 title = "디버그";
-                message = "다음날로 이동하시겠습니까?\n이 버튼은 디버그 및 과제평과를 위해 만들었습니다.\n\'이동\'버튼을 누르면 어플이 종료됩니다. 다시 시작해주세요.";
+                message = "다음날로 이동하시겠습니까?\n이 버튼은 디버그 및 과제평가를 위해 만들었습니다.\n\'이동\'버튼을 누르면 어플이 종료됩니다. 다시 시작해주세요.";
                 pb = "이동";
                 nb = "취소";
                 listenerPositive = new DialogInterface.OnClickListener() {
@@ -296,7 +294,7 @@ public class MainActivity extends Activity {
 
         TabHost.TabSpec tab2 = tabhost.newTabSpec("Tab Spac 2");
         tab2.setContent(R.id.tab2);
-        tab2.setIndicator("passed");
+        tab2.setIndicator("기록");
         tabhost.addTab(tab2);
 
         TabHost.TabSpec tab3 = tabhost.newTabSpec("Tab Spac 3");
